Add files via upload
About the Project
This is a self-paced lab that takes place in the Google Cloud console. In this lab, I learnt how to use Google's Vertex Al SDK to interact with the powerful Gemini generative Al model, enabling me to send text based chat prompts as an input and receive personalized streaming and non-streaming chat responses.